+++ higlass/server.py
@@ -1,4 +1,5 @@
 from functools import partial
+from itertools import count
 from io import StringIO
 import multiprocess as mp
 import cytoolz as toolz
@@ -11,6 +12,7 @@ import tempfile
 import json
 import time
 import os
+import atexit
 
 from flask import Flask
 from flask import request, jsonify
@@ -19,6 +21,8 @@ from flask import request, jsonify
 from flask_cors import CORS
 
 import requests
+import requests_unixsocket
+from urllib.request import quote
 import slugid
 import sh
 
@@ -37,6 +41,7 @@ log.disabled = True
 # The following line is also needed to turn off all debug logs
 os.environ["WERKZEUG_RUN_MAIN"] = "true"
 
+requests_unixsocket.monkeypatch()
 
 def create_app(name, tilesets, fuse=None):
     app = Flask(name)
@@ -205,6 +210,23 @@ class ServerError(Exception):
     pass
 
 
+def get_free_socket(sock_dir):
+    for p in count():
+        sock_file = op.join(sock_dir, str(p))
+        s = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+        try:
+            # Check if socket exists and alive
+            s.connect(sock_file)
+        except OSError:
+            # Socket is missing or dead, use it
+            try:
+                os.remove(sock_file)
+            except FileNotFoundError:
+                pass
+            return p
+        # Socket is alive, move on
+        s.close()
+
 def get_open_port():
     s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
     s.bind(("", 0))
@@ -321,8 +343,11 @@ class Server:
         A list of tilesets to serve (see higlass.tilesets)
     host : str, optional
         The host this server is running on. Usually just localhost.
+        If starts with unix://, then it's treated as a unix socket. If also ends with a '/',
+        then it's treated as a directory of unix sockets, with port being the file name.
     port : int, optional
-        The port that this server will run on.
+        The port that this server will run on. If host starts with unix://, then this is the
+        file name for the socket inside the base directory.
     name : str, optional
         A name for the Flask app being served. If not provided, a
         unique name will be generated. The app's logger inherits this
@@ -339,8 +364,8 @@ class Server:
         StringIO stream in memory.
     root_api_address: str, optional
         Root path of the reported api_address, instead of http://{host}:{port}.
-        Can be used for redirecting clients to a proxy. Can include {host} and {port}
-        to use these variables.
+        Can be used for redirecting clients to a proxy. Can include {host}, {port}
+        and {unix_filename} to use these variables.
 
     """
 
@@ -401,8 +426,20 @@ class Server:
             self.processes[puid].terminate()
             del self.processes[puid]
 
-        if self.port is None:
-            self.port = get_open_port()
+        self._use_unix = self.host.startswith("unix://")
+        if not self._use_unix:
+            if self.port is None:
+                self.port = get_open_port()
+            self._unix_filename = None
+        else:
+            host_no_schema = self.host[7:]
+            if host_no_schema[-1] == '/':
+                os.makedirs(host_no_schema, mode=0o700, exist_ok=True)
+            if op.isdir(host_no_schema):
+                if self.port is None:
+                    self.port = get_free_socket(host_no_schema)
+                self.host = op.join(self.host, str(self.port))
+            self._unix_filename = self.host[7:]
 
         # we're going to assign a uuid to each server process so that if
         # anything goes wrong, the variable referencing the process doesn't get
@@ -418,18 +455,25 @@ class Server:
             use_reloader=False,
             **kwargs
         )
+        atexit.register(self._unix_cleanup)
         self.processes[uuid] = mp.Process(target=target)
         self.processes[uuid].start()
 
         self.connected = False
         while not self.connected:
             try:
-                url = "http://{}:{}/api/v1".format(self.host, self.port)
+                url = self._get_url('api/v1')
                 r = requests.head(url)
                 if r.ok:
                     self.connected = True
-            except requests.ConnectionError:
+            except requests.ConnectionError as e:
                 time.sleep(0.2)
+
+    def _get_url(self, path):
+        if self._use_unix:
+            host = quote(self._unix_filename, safe='')
+            return "http+unix://{}/{}".format(host, path)
+        return "http://{}:{}/{}".format(self.host, str(self.port), path)
 
     def stop(self):
         """
@@ -439,14 +483,20 @@ class Server:
             self.fuse_process.teardown()
         for uuid in self.processes:
             self.processes[uuid].terminate()
+        self._unix_cleanup()
+
+    def _unix_cleanup(self):
+        if self._use_unix:
+            try:
+                os.remove(self._unix_filename)
+            except FileNotFoundError:
+                pass
 
     def tileset_info(self, uid):
         """
         Return the tileset info for the given tileset
         """
-        url = "http://{host}:{port}/api/v1/tileset_info/?d={uid}".format(
-            host=self.host, port=self.port, uid=uid
-        )
+        url = self._get_url("api/v1/tileset_info/?d={uid}".format(uid=uid))
 
         req = requests.get(url)
         if req.status_code != 200:
@@ -463,9 +513,7 @@ class Server:
         tile_id = "{uid}.{z}.{x}".format(uid=uid, z=z, x=x)
         if y is not None:
             tile_id += ".{y}".format(y=y)
-        url = "http://{host}:{port}/api/v1/tiles/?d={tile_id}".format(
-            host=self.host, port=self.port, tile_id=tile_id
-        )
+        url = self._get_url("api/v1/tiles/?d={tile_id}".format(tile_id=tile_id))
 
         req = requests.get(url)
         if req.status_code != 200:
@@ -478,9 +526,7 @@ class Server:
         """
         Return the chromosome sizes from the given filename
         """
-        url = "http://{host}:{port}/api/v1/chrom-sizes/?id={uid}".format(
-            host=self.host, port=self.port, uid=uid
-        )
+        url = self._get_url("api/v1/chrom-sizes/?id={uid}".format(uid=uid))
 
         req = requests.get(url)
         if req.status_code != 200:
@@ -491,6 +537,6 @@ class Server:
     @property
     def api_address(self):
         if self._root_api_address is None:
-            return "http://{host}:{port}/api/v1".format(host=self.host, port=self.port)
-        root = self._root_api_address.format(host=self.host, port=self.port)
+            return self._get_url('')
+        root = self._root_api_address.format(host=self.host, port=self.port, unix_filename=self._unix_filename or '')
         return "{}/api/v1".format(root)